murrasil/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
feedparser==6.0.10
openai==1.10.0
apscheduler==3.10.4
python-dotenv==1.0.0
aiohttp==3.9.1
